Move TargetSet and Mode into own modules
