+++ aquadoggo/src/replication/mode.rs
@@ -1,0 +1,72 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use std::fmt::{self, Display};
+
+use p2panda_rs::Human;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum Mode {
+    Naive,
+    SetReconciliation,
+    Unknown,
+}
+
+impl Mode {
+    pub fn as_str(&self) -> &str {
+        match self {
+            Mode::Naive => "naive",
+            Mode::SetReconciliation => "set-reconciliation",
+            Mode::Unknown => "unknown",
+        }
+    }
+
+    pub fn as_u64(&self) -> u64 {
+        match self {
+            Mode::Naive => 0,
+            Mode::SetReconciliation => 1,
+            Mode::Unknown => unreachable!("Can't create an unknown replication mode"),
+        }
+    }
+}
+
+impl From<u64> for Mode {
+    fn from(value: u64) -> Self {
+        match value {
+            0 => Mode::Naive,
+            1 => Mode::SetReconciliation,
+            _ => Mode::Unknown,
+        }
+    }
+}
+
+impl Serialize for Mode {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_u64(self.as_u64())
+    }
+}
+
+impl<'de> Deserialize<'de> for Mode {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let mode = u64::deserialize(deserializer)?;
+        Ok(mode.into())
+    }
+}
+
+impl Display for Mode {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.as_u64())
+    }
+}
+
+impl Human for Mode {
+    fn display(&self) -> String {
+        self.as_str().to_owned()
+    }
+}
